+++ packages/core-components/src/lib/atoms/index.js
@@ -1,5 +1,5 @@
 export * from './accordion';
-export { default as Alert } from './alert/Alert.svelte';
+export * from './alert';
 export * from './skeletons';
 export * from './button';
 export { default as CopyButton } from './copy-button/CopyButton.svelte';