+++ packages/core-components/src/lib/unsorted/viz/map/_USMap.svelte
@@ -5,7 +5,7 @@
 <script>
 	import EChartsMap from './EChartsMap.svelte';
 	import ErrorChart from '../core/ErrorChart.svelte';
-	import { strictBuild } from '../context';
+	import { strictBuild } from '@evidence-dev/component-utilities/chartContext';
 	import checkInputs from '@evidence-dev/component-utilities/checkInputs';
 	import formatTitle from '@evidence-dev/component-utilities/formatTitle';
 	import getColumnSummary from '@evidence-dev/component-utilities/getColumnSummary';
@@ -17,6 +17,7 @@
 	import InvisibleLinks from '../../../atoms/InvisibleLinks.svelte';
 
 	export let data = undefined;
+	export let queryID = undefined;
 
 	export let state = undefined;
 	export let value = undefined;
@@ -298,7 +299,15 @@
 </script>
 
 {#if !error}
-	<EChartsMap {extraHeight} {config} {data} {hasLink} {echartsOptions} {printEchartsConfig} />
+	<EChartsMap
+		{extraHeight}
+		{config}
+		{data}
+		{queryID}
+		{hasLink}
+		{echartsOptions}
+		{printEchartsConfig}
+	/>
 
 	{#if link}
 		<InvisibleLinks {data} {link} />