+++ packages/query-store/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @evidence-dev/query-store
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [e4e7d822]
+  - @evidence-dev/universal-sql@2.0.2
+
 ## 2.0.1
 
 ### Patch Changes